% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{crSearch}
\alias{crSearch}
\title{Function to perform crossref search on a query term for a given time window}
\usage{
crSearch(
  query,
  since = as.character(Sys.Date() - 7),
  until = as.character(Sys.Date() + 365),
  relevant = ""
)
}
\arguments{
\item{query}{Query term}

\item{since}{Date articles published since (default is today minus a week)}

\item{until}{Maximum date (default is today plus 1 year)}

\item{relevant}{string of terms (separated by |) to insist on in title}
}
\value{
A tibble of article metadata, unless over 1000 articles are returned in 
which case it asks the user to modify their search
}
\description{
Some records will return an abstract but coverage will be vastly improved
by running getAbstract() where they are missing
}
