% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{fetchFromDOI}
\alias{fetchFromDOI}
\title{Retrieve info from DOI}
\usage{
fetchFromDOI(articledoi)
}
\arguments{
\item{articledoi}{DOI of the article}
}
\value{
A tibble with all the info that could be retrieved from crossref
}
\description{
Use crossref to retrieve information about an article via crossref from its DOI
}
