% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{fetchFromTitle}
\alias{fetchFromTitle}
\title{Retrieve DOI from title}
\usage{
fetchFromTitle(articletitle)
}
\arguments{
\item{articletitle}{title of the article}
}
\value{
A tibble with all the info that could be retrieved from crossref
}
\description{
Use crossref to retrieve information about an article via crossref from its title
}
