% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{getAbstract}
\alias{getAbstract}
\title{Get an abstract}
\usage{
getAbstract(articledoi)
}
\arguments{
\item{articledoi}{DOI identifier of the article}
}
\value{
a tibble with DOI and abstract
}
\description{
Retrieve an abstract from a DOI (searches pubmed). Use slowly(getAbstract, rate = purrr::rate_delay(pause = 1)) if used in a map loop.
}
