% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{getBody}
\alias{getBody}
\title{Get the body of an article}
\usage{
getBody(articledoi)
}
\arguments{
\item{articledoi}{DOI identifier of the article}
}
\value{
a tibble with DOI and body
}
\description{
Retrieve the full text of an article from a DOI, where access is available
Limited to first 30000 characters to write to CSV
}
