% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emails.R
\name{getDOIs}
\alias{getDOIs}
\title{Extract list of DOIs from a vector of strings}
\usage{
getDOIs(stringlist)
}
\arguments{
\item{stringlist}{input vector of strings}
}
\value{
a vector of strings that are all DOIs
}
\description{
Extract list of DOIs from a vector of strings
}
