% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emails.R
\name{getLinkText}
\alias{getLinkText}
\title{Get link text from an email alert}
\usage{
getLinkText(emailfolder, email)
}
\arguments{
\item{emailfolder}{Path to folder containing email}

\item{email}{filename of email}
}
\value{
A vector of text strings
}
\description{
Function to return all link text in an email
}
