% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emails.R
\name{getTitle}
\alias{getTitle}
\title{Extract list of titles (non-DOIs) from a vector of strings}
\usage{
getTitle(stringlist)
}
\arguments{
\item{stringlist}{input vector of strings}
}
\value{
a vector of strings that are not DOIs
}
\description{
Extract list of titles (non-DOIs) from a vector of strings
}
