% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{maketable}
\alias{maketable}
\title{Make a lemma table}
\usage{
maketable(data, col)
}
\arguments{
\item{data}{The dataset}

\item{col}{A string with the column you want the lemma table for}
}
\value{
A count table of lemmas by pos
}
\description{
Make a lemma table
}
