% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\name{makewordcloud}
\alias{makewordcloud}
\title{Make a wordcloud}
\usage{
makewordcloud(wordtable, wordvar, size = 4)
}
\arguments{
\item{wordtable}{A table with words and counts (n)}

\item{wordvar}{The variable with the words}

\item{size}{Upper limit of word size}
}
\description{
Make a wordcloud
}
