% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_dataquality.R
\name{mod_dataquality_ui}
\alias{mod_dataquality_ui}
\alias{mod_dataquality_server}
\title{mod_dataquality_ui and mod_dataquality_server}
\usage{
mod_dataquality_ui(id)

mod_dataquality_server(input, output, session, data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
