% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tables.R
\name{mod_tables_ui}
\alias{mod_tables_ui}
\alias{mod_tables_server}
\title{mod_tables_ui and mod_tables_server}
\usage{
mod_tables_ui(id)

mod_tables_server(input, output, session, data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
