% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_timetrends.R
\name{mod_timetrends_ui}
\alias{mod_timetrends_ui}
\alias{mod_timetrends_server}
\title{mod_timetrends_ui and mod_timetrends_server}
\usage{
mod_timetrends_ui(id)

mod_timetrends_server(input, output, session, data, data2)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
