% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_toplist.R
\name{mod_toplist_ui}
\alias{mod_toplist_ui}
\alias{mod_toplist_server}
\title{mod_toplist_ui and mod_toplist_server}
\usage{
mod_toplist_ui(id)

mod_toplist_server(input, output, session, data, var)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
