% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_wordfreq.R
\name{mod_wordfreq_ui}
\alias{mod_wordfreq_ui}
\alias{mod_wordfreq_server}
\title{mod_wordfreq_ui and mod_wordfreq_server}
\usage{
mod_wordfreq_ui(id)

mod_wordfreq_server(input, output, session, data)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
