% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{plosSearch}
\alias{plosSearch}
\title{Searches for a term in PLOS since a given publication date}
\usage{
plosSearch(query, since = Sys.Date() - 7, limit = 100)
}
\arguments{
\item{query}{text of the query}

\item{since}{date, default is today minus a week}

\item{limit}{maximum number of articles to return (default = 100)}
}
\description{
Searches for a term in PLOS since a given publication date
}
\examples{
# search for just the word "microbiome" in all fields
plosSearch("microbiome")
# search for the term "gut microbiome"
plosSearch("gut microbiome")
# search for the words "gut" and "microbiome" within 10 words of each other
plosSearch("'gut microbiome'~10")

}
